Cleanup, split classes into files
